server/pyproject.toml
[project]
name = "spelling-segmenter"
version = "1.0.0"
description = "Text segmentation service for spelling practice app"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "spacy>=3.8.4",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
    "openai>=1.57.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app", "segmenters"]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
